+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- For the IPerson behavior, compute the title from first and last name and add
+  title (not required, hidden) and description to the IPerson behavior.
+  [thet]
 
 
 1.1 (2015-03-04)